Seperate database, password manager and CLI logic into seperate modules.
